fix(use-workspace): correct logic for parsing workspace URL
This commit fixes a critical logical error in `getWorkspaceLocationByEnv` function within `use-workspace.ts`. Previously, the condition mistakenly checked if a workspace location URL can be parsed, and if so, it incorrectly returned null. This behavior is opposite of what should happen: we should return null if the URL *cannot* be parsed. This patch corrects this by adding a missing negation operator (!) in the if-statement condition. This change ensures that when an environment variable `WORKSPACE_LOCATION` is set but cannot be correctly parsed into a valid URL with respect to the current working directory, the function will rightly return null as intended. This is important to prevent runtime errors when attempting to use malformed URLs as workspace locations.