feat: add template copying functionality to useWorkspace
This commit introduces the ability to specify a template directory when creating a new workspace with the useWorkspace function. This feature allows the user to initialize a workspace with a predefined set of files and directory structure. The option 'template' can be passed to useWorkspace, which should be a URL pointing to the directory containing the template files. When specified, all files within the template directory will be recursively copied into the newly created workspace before any further operations are performed. This enhancement is particularly useful for setting up consistent testing or development environments quickly and efficiently. In addition to this, a new file 'foo' with content 'biz' was added to the `__examples__/sample1` directory to serve as a part of the template in testing scenarios. Modifications were also made to the test suites to support and validate this new functionality.