chore: bump package version to 0.1.4 for new features
This commit increases the package version from 0.1.3 to 0.1.4. The version bump is necessary to reflect the addition of new features and improvements to the 'use-workspace' package. This update ensures that consumers of the package have access to the latest functionalities and enhancements. Users are encouraged to update to this latest version to benefit from the improvements.